requirements.txt
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
requests==2.32.3
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
langdetect==1.0.9
nltk==3.8.1
python-magic==0.4.27
tqdm==4.66.2
